Moved Requirements to separate component.
