+++ src/modules/Preservation/views/ScopeOverview/TagFlyout/PreservationTab/PreservationTab.style.ts
@@ -1,4 +1,4 @@
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import { tokens } from '@equinor/eds-tokens';
 
 export const Container = styled.div`
@@ -16,25 +16,6 @@ export const Details = styled.div`
     display: flex;
     flex-direction: column;
     padding: calc(var(--grid-unit) * 2);
-`;
-
-export const RequirementContainer = styled.div`
-    border-bottom: 1px solid ${tokens.colors.ui.background__medium.rgba};
-    padding: 
-        calc(var(--grid-unit) * 4) 
-        calc(var(--grid-unit) * 2) 
-        calc(var(--grid-unit) * 2) 
-        calc(var(--grid-unit) * 2);
-`;
-
-export const RequirementSection = styled.div`
-    display: flex;
-    flex-direction: column;
-    margin-bottom: calc(var(--grid-unit) * 2);
-`;
-
-export const Field = styled.div`
-    margin-bottom: var(--grid-unit);
 `;
 
 export const RemarkContainer = styled.div`
@@ -56,10 +37,4 @@ export const GridSecondRow = styled.div`
     grid-template-rows: auto auto;
     grid-column-gap: calc(var(--grid-unit) * 3);
     float: left;
-`;
-
-export const NextInfo = styled.span<{ isOverdue: boolean }>`
-    ${(props): any => props.isOverdue && css`
-        color: ${tokens.colors.interactive.danger__text.rgba};
-    `}
 `;
